+++ src/server/plugins/engine/types.ts
@@ -4,6 +4,7 @@ import {
   type List,
   type Page
 } from '@defra/forms-model'
+import { type PluginProperties } from '@hapi/hapi'
 import { type JoiExpression, type ValidationErrorItem } from 'joi'
 
 import { FormComponent } from '~/src/server/plugins/engine/components/FormComponent.js'
@@ -13,10 +14,12 @@ import {
   type ComponentText,
   type ComponentViewModel
 } from '~/src/server/plugins/engine/components/types.js'
+import { type FormModel } from '~/src/server/plugins/engine/models/index.js'
 import { type PageController } from '~/src/server/plugins/engine/pageControllers/PageController.js'
 import { type PageControllerClass } from '~/src/server/plugins/engine/pageControllers/helpers.js'
 import { type ViewContext } from '~/src/server/plugins/nunjucks/types.js'
 import { type FormAction, type FormRequest } from '~/src/server/routes/types.js'
+import { type Services } from '~/src/server/types.js'
 
 /**
  * Form submission state stores the following in Redis:
@@ -324,4 +327,18 @@ export interface ErrorMessageTemplate {
 export interface ErrorMessageTemplateList {
   baseErrors: ErrorMessageTemplate[]
   advancedSettingsErrors: ErrorMessageTemplate[]
+}
+
+export interface PluginOptions {
+  model?: FormModel
+  services?: Services
+  controllers?: Record<string, typeof PageController>
+  cacheName?: string
+  filters?: Record<string, FilterFunction>
+  pluginPath?: string
+  nunjucks: {
+    baseLayoutPath: string
+    paths: string[]
+  }
+  viewContext: PluginProperties['forms-engine-plugin']['viewContext']
 }
